refactor(coding-agents): extract the pi extension adapter Prime Agent shares
Prime Agent is a fork of pi and its extension API is pi's: the same
`extensions` array in settings.json, the same `before_agent_start` /
`agent_end` events, the same `registerTool`. The adapter for it was written as
if that API were Prime Agent's own, so a second host on the same API had
nowhere to plug in.

Move it to harness/pi-extension.ts behind `createPiExtension(harness)`,
mirroring how harness/plugin-entry.ts already serves the opencode family, and
leave prime-agent.ts as a thin entrypoint that names its harness. The message
normalizer moves with it: the shape it parses is pi's, not the fork's, so
transcript-prime-agent.ts becomes transcript-pi.ts (PiMessage, readPiMessages).

No behaviour change — the harness is still "prime-agent", and the adapter tests
move across unchanged apart from being parameterised by harness name.
